pyproject.toml
[project]
name = "client1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "mcp>=1.4.1",
    "python-dotenv>=1.0.1",
]
